Prevent loading BBox from frames that are not available as pointcloud

+++ lidarFrameManager.h
@@ -19,7 +19,7 @@ public:
 
   double getFrameTimestamp(unsigned int frameId) const;
 
-  unsigned int getNbFrames() const {
+  size_t getNbFrames() const {
     return m_frames.size();
   }
